pyproject.toml
[tool.poetry]
name = "fastapi-service"
version = "0.1.0"
description = ""
authors = ["joebui <dienbuijoe@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
psycopg = { extras = ["binary", "pool"], version = "^3.1.19" }
pq = "^1.9.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
structlog = "^24.2.0"

[tool.poetry.dev-dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 60

[tool.isort]
profile = "black"
virtual_env = "env"
